chore: 회원가입 닉네임 중복확인 로직 AuthController쪽으로 이동
