+++ src/main/java/com/kusitms/jipbap/user/UserService.java
@@ -48,9 +48,5 @@ public class UserService {
         user.updateRefreshToken(null);
     }
 
-    public String checkNicknameIsDuplicate(String nickname){
-        if(userRepository.existsByUsername(nickname)) throw new UsernameExistsException("이미 존재하는 닉네임입니다.");
-        return "사용 가능한 닉네임입니다.";
-    }
 }
